requirements.txt
alabaster==0.7.12
alembic==0.8.10
anaconda-client==1.7.2
anaconda-project==0.8.2
apache-airflow==1.10.1
apache-beam==2.9.0
appdirs==1.4.3
appnope==0.1.0
appscript==1.0.1
asn1crypto==0.24.0
astroid==1.6.5
astropy==2.0.9
atomicwrites==1.2.1
attrs==18.2.0
Automat==0.7.0
avro==1.8.2
Babel==2.6.0
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.6.3
bitarray==0.8.3
bkcharts==0.2
blaze==0.11.3
bleach==2.1.4
bokeh==1.0.2
boto==2.49.0
Bottleneck==1.2.1
cachetools==3.0.0
cdecimal==2.3
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
clyent==1.2.2
colorama==0.4.0
configparser==3.5.0
constantly==15.1.0
contextlib2==0.5.5
crcmod==1.7
croniter==0.3.26
cryptography==2.4.2
cx-Oracle==7.0.0
cycler==0.10.0
Cython==0.29
cytoolz==0.9.0.1
dask==1.0.0
datashape==0.5.4
decorator==4.3.0
defusedxml==0.5.0
dill==0.2.8.2
distributed==1.25.0
docopt==0.6.2
docutils==0.14
entrypoints==0.2.3
enum34==1.1.6
et-xmlfile==1.0.1
fastavro==0.21.16
fastcache==1.0.2
fasteners==0.14.1
filelock==3.0.10
Flask==0.12.4
Flask-Admin==1.4.1
Flask-AppBuilder==1.12.2
Flask-Babel==0.12.2
Flask-Caching==1.3.3
Flask-Cors==3.0.7
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.3.2
flask-swagger==0.2.13
Flask-WTF==0.14.2
funcsigs==1.0.0
functools32==3.2.3.post2
future==0.16.0
futures==3.2.0
gevent==1.3.7
gitdb2==2.0.5
GitPython==2.1.11
glob2==0.6
gmpy2==2.0.8
google-api-core==1.7.0
google-apitools==0.5.24
google-auth==1.6.2
google-cloud-bigquery==1.6.1
google-cloud-core==0.29.1
google-cloud-pubsub==0.35.4
google-resumable-media==0.3.2
googleapis-common-protos==1.5.5
googledatastore==7.0.2
greenlet==0.4.15
grin==1.2.1
grpc-google-iam-v1==0.11.4
grpcio==1.17.1
gunicorn==19.9.0
h5py==2.8.0
hdfs==2.2.2
heapdict==1.0.0
html5lib==1.0.1
httplib2==0.11.3
hyperlink==18.0.0
idna==2.7
imageio==2.4.1
imagesize==1.1.0
importlib-metadata==0.6
incremental==17.5.0
ipaddress==1.0.22
ipykernel==4.10.0
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
iso8601==0.1.12
isort==4.3.4
itsdangerous==1.1.0
jdcal==1.4
jedi==0.13.1
Jinja2==2.8.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.33.11
jupyterlab-launcher==0.11.2
keyring==16.1.1
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
linecache2==1.0.0
llvmlite==0.26.0
locket==0.2.0
lockfile==0.12.2
lxml==4.2.5
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.1.0
matplotlib==2.2.3
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.6
mkl-random==1.0.1
mock==2.0.0
monotonic==1.5
more-itertools==4.3.0
mpmath==1.0.0
msgpack==0.5.6
multipledispatch==0.6.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.2
nltk==3.3
nose==1.3.7
notebook==5.7.2
numba==0.41.0+0.gf118cda06.dirty
numexpr==2.6.8
numpy==1.15.4
numpydoc==0.8.0
oauth2client==3.0.0
odo==0.5.1
olefile==0.46
openpyxl==2.5.11
ordereddict==1.1
packaging==18.0
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
partd==0.3.9
path.py==11.5.0
pathlib2==2.3.2
patsy==0.5.1
pbr==5.1.1
pendulum==1.4.4
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
pkginfo==1.4.2
pluggy==0.8.0
ply==3.11
prometheus-client==0.4.2
prompt-toolkit==1.0.15
proto-google-cloud-datastore-v1==0.90.4
protobuf==3.6.1
psutil==4.4.2
ptyprocess==0.6.0
py==1.7.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
PyAudio==0.2.11
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.2
pydot==1.2.4
pyflakes==2.0.0
Pygments==2.2.0
pylint==1.9.2
pyodbc==4.0.24
pyOpenSSL==18.0.0
pyparsing==2.3.0
PySocks==1.6.8
pyspark==2.1.0+hadoop2.7
pytest==4.0.1
python-daemon==2.1.2
python-dateutil==2.7.5
python-editor==1.0.3
python-nvd3==0.15.0
python-openid==2.2.5
python-slugify==2.0.0
pytz==2018.4
pytzdata==2018.7
PyVCF==0.6.8
PyWavelets==1.0.1
PyYAML==3.13
pyzmq==17.1.2
QtAwesome==0.5.3
qtconsole==4.4.2
QtPy==1.5.2
requests==2.20.1
rope==0.11.0
rsa==4.0
ruamel-yaml==0.15.46
scandir==1.9.0
scikit-image==0.14.0
scikit-learn==0.20.1
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
service-identity==17.0.0
setproctitle==1.1.10
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
smmap2==2.0.5
snowballstemmer==1.2.1
sortedcollections==1.0.1
sortedcontainers==2.1.0
Sphinx==1.8.2
sphinx-rtd-theme==0.4.2
sphinxcontrib-websupport==1.1.0
spyder==3.3.2
spyder-kernels==0.3.0
SQLAlchemy==1.1.18
statsmodels==0.9.0
subprocess32==3.5.3
sympy==1.3
tables==3.4.4
tabulate==0.8.2
tblib==1.3.2
tenacity==4.8.0
terminado==0.8.1
testpath==0.4.2
thrift==0.11.0
toolz==0.9.0
tornado==5.1.1
tqdm==4.28.1
traceback2==1.4.0
traitlets==4.3.2
Twisted==17.5.0
typing==3.6.6
tzlocal==1.5.1
unicodecsv==0.14.1
Unidecode==1.0.23
unittest2==1.1.0
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.10.11
WTForms==2.2.1
wurlitzer==1.0.2
xlrd==1.1.0
XlsxWriter==1.1.2
xlwings==0.15.0
xlwt==1.2.0
zict==0.1.3
zope.deprecation==4.4.0
zope.interface==4.6.0
